Password Vlidation Continued
I have implemented a few new features (one of them I took from the internet)

The app now checks password, length, makes sure the password has at least one letter and one special character, chceks if entered password is one from the 10 million password txt file [from github], and a last but not least: checks if the password is similar to the common password by using an algorithim which I took from the internet.

Last check, the one with the algorithm freezes the app for couple of seconds, I guess this freeze will be longer on low spec machines, but for now I will move on and fix this issue later.